requirements.txt
python-telegram-bot
youtube-search-python
yt_dlp
aiohttp